Use XOR for equality for massive performance boost
This is based on Andreas's suggestion in GHC #17464. By using xorI#
instead of eqChar#, we can thwart GHC's eagerness to convert equality
tests into casing in Core. Further down the compilation pipeline, this
prevents GHC from producing assembly with a bunch of needless branching.
For reasons I have not really dug into, the performance boost here is
massive, much larger than expected. In the Traffic-Local-A microbenchmark,
the run time drops from 2.4us to 1.4us.

This also fixes a minor issue where the Service field was mislabeled.
This change is unrelated to the improvement in performance.